Add files via upload
The data used in this study is from the 2015 American Community 5 year Survey (2011 - 2015). We can use this data to explore the evolution of post-recession unemployment in the US. I am particularly curious to see if race has an correlation to unemployment (per state) in the US, and if so, how strong is this correlation?